requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# API & Web Scraping
requests>=2.31.0
python-dotenv>=1.0.0

# NLP & Text Processing
nltk>=3.8.0
scikit-learn>=1.3.0
textblob>=0.17.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
flask>=2.3.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
notebook>=6.5.4

# Geospatial (optional for geo-visualization)
geopandas>=0.13.0
folium>=0.14.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
